New energy net! We're using ConductorSet now.
set buckets. "Disjoint sets" kiinda not really at all?
New charges create an empty set containing only themselves, if
they can't join a neighboring set. Unfortunately TEs don't know
what the world is when they're loading, so this will cause an
unnecessary set to be created for each conductor when the world
loads. This'll be performance bad, but behaviorly is mostly
taken care of by having sets merge together when they're new,
and also randomly every 15 (next commit: 30) seconds per
conductor. The creating-a-ton-of-sets-issue might be resolved
by just not creating sets when loading the Charge. This might
require more checks in some other functions, as it is possible
that they'll be looked at within the first tick. Also we'd need
to keep track of their charge in the mean time.

Small sets will join up into larger conductor sets. They also
keep track of their neighbors. The maximum size of a set is 24.

Charge is stored in the ConductorSet, and loaded dynamically
from/to Charge objects as they request it.

Probable bug that might be fixed in next commit:
Sets that get half unloaded will duplicate the energy that got
saved into the chunk. Fixed by removing invalidated TEs charge.