Solar Turbines are being replaced with Solar Boilers and Steam Turbines.
The solar boiler is setup the same way the old turbine was. Instead of
sucking water up from the bottom, it can also be piped in using some
other mod's fluid transport. The steam comes out of the top side.

Solar boilers can also be stacked on top of eachother. Steam will be
sent to the topmost boiler, and water will be shared. Steam will be sent
to a Steam Turbine if it's placed ontop of the Boiler.

This system is compatible with Railcraft steam.